+++ packages/plugin-link/web/src/LinkViewer.jsx
@@ -20,6 +20,7 @@ class LinkViewer extends Component {
     children: PropTypes.node,
     anchorTarget: PropTypes.string,
     relValue: PropTypes.string,
+    customAnchorScroll: PropTypes.func,
     settings: PropTypes.object,
     isInEditor: PropTypes.bool,
     config: PropTypes.object,
@@ -40,7 +41,7 @@ class LinkViewer extends Component {
   }
 
   handleClick = event => {
-    const { componentData, isInEditor, config, helpers } = this.props;
+    const { componentData, isInEditor, config, helpers, customAnchorScroll } = this.props;
     const settings = config?.[LINK_TYPE];
     if (settings) {
       const { onClick } = settings;
@@ -51,10 +52,14 @@ class LinkViewer extends Component {
         event.stopPropagation(); // fix problem with wix platform, where it wouldn't scroll and sometimes jump to different page
         if (!isInEditor) {
           event.preventDefault();
-          const anchorString = `viewer-${anchor}`;
-          const element = document.getElementById(anchorString);
-          addAnchorTagToUrl(anchorString);
-          anchorScroll(element);
+          if (customAnchorScroll) {
+            customAnchorScroll(event, anchor);
+          } else {
+            const anchorString = `viewer-${anchor}`;
+            const element = document.getElementById(anchorString);
+            addAnchorTagToUrl(anchorString);
+            anchorScroll(element);
+          }
         }
       }
     }