+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -10,6 +10,7 @@ import {
   EditorPlugin,
   ViewerPlugin,
   onAtomicBlockFocus,
+  CustomAnchorScroll,
 } from 'wix-rich-content-common';
 import { EditorState, EditorProps } from 'draft-js';
 import { PreviewConfig } from 'wix-rich-content-preview';
@@ -126,6 +127,7 @@ export interface LinkSettings {
   anchorTarget?: AnchorTarget;
   relValue?: RelValue;
   rel?: Link_Rel;
+  customAnchorScroll?: CustomAnchorScroll;
 }
 
 export type TextAlignment = 'left' | 'right';